@@ -3,24 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.1",
+    "@mui/material": "^7.3.1",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "bootstrap": "^4.6.0",
     "ityped": "^1.0.3",
     "jquery": "^3.6.0",
-    "material-ui": "^0.20.2",
-    "netlify-cli": "^3.33.2",
-    "node-sass": "^6.0.0",
+    "netlify-cli": "^23.1.4",
+    "node-sass": "^9.0.0",
     "popper.js": "^1.16.1",
-    "react": "^17.0.2",
+    "react": "^19.1.1",
     "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^19.1.1",
     "react-fade-in": "^2.0.1",
     "react-router-dom": "^5.0.1",
-    "react-scripts": "4.0.3"
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -47,7 +48,12 @@
     ]
   },
   "devDependencies": {
-    "node-sass": "^6.0.0",
-    "sass-loader": "^11.1.1"
+    "nan": "2.18.0",
+    "node-gyp": "11.3.0",
+    "node-sass": "^9.0.0",
+    "sass-loader": "^16.0.5"
+  },
+  "resolutions": {
+    "nan": "2.18.0"
   }
 }
